+++ clients/camunda-spring-boot-starter/src/main/java/io/camunda/client/jobhandling/JobWorkerChangeSet.java
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.spring.client.jobhandling;
+package io.camunda.client.jobhandling;
 
-import io.camunda.spring.client.annotation.value.JobWorkerValue;
-import io.camunda.spring.client.annotation.value.JobWorkerValue.FetchVariable;
-import io.camunda.spring.client.annotation.value.JobWorkerValue.FieldSource;
-import io.camunda.spring.client.annotation.value.JobWorkerValue.Name;
-import io.camunda.spring.client.annotation.value.JobWorkerValue.Type;
+import io.camunda.client.annotation.value.JobWorkerValue;
+import io.camunda.client.annotation.value.JobWorkerValue.FetchVariable;
+import io.camunda.client.annotation.value.JobWorkerValue.FieldSource;
+import io.camunda.client.annotation.value.JobWorkerValue.Name;
+import io.camunda.client.annotation.value.JobWorkerValue.Type;
 import java.time.Duration;
 import java.util.List;
 import java.util.Objects;
