+++ clients/camunda-spring-boot-starter/src/main/java/io/camunda/client/jobhandling/JobWorkerManager.java
@@ -16,27 +16,12 @@
 package io.camunda.client.jobhandling;
 
 import io.camunda.client.CamundaClient;
+import io.camunda.client.annotation.customizer.JobWorkerValueCustomizer;
 import io.camunda.client.annotation.value.JobWorkerValue;
-import io.camunda.client.api.worker.BackoffSupplier;
-import io.camunda.client.api.worker.JobHandler;
 import io.camunda.client.api.worker.JobWorker;
-import io.camunda.spring.client.annotation.customizer.JobWorkerValueCustomizer;
-import io.camunda.spring.client.annotation.value.JobWorkerValue;
-import io.camunda.spring.client.jobhandling.JobWorkerChangeSet.CreateChangeSet;
-import io.camunda.spring.client.jobhandling.JobWorkerChangeSet.EnabledChangeSet;
-import io.camunda.spring.client.jobhandling.JobWorkerChangeSet.ResetChangeSet;
-import io.camunda.spring.client.jobhandling.parameter.ParameterResolverStrategy;
-import io.camunda.spring.client.jobhandling.result.ResultProcessorStrategy;
-import io.camunda.spring.client.metrics.MetricsRecorder;
-import io.camunda.client.jobhandling.JobWorkerChangeSet.NoopChangeSet;
+import io.camunda.client.jobhandling.JobWorkerChangeSet.CreateChangeSet;
+import io.camunda.client.jobhandling.JobWorkerChangeSet.EnabledChangeSet;
 import io.camunda.client.jobhandling.JobWorkerChangeSet.ResetChangeSet;
-import io.camunda.client.jobhandling.JobExceptionHandlerSupplier.JobExceptionHandlerSupplierContext;
-import io.camunda.client.jobhandling.parameter.ParameterResolver;
-import io.camunda.client.jobhandling.parameter.ParameterResolverStrategy;
-import io.camunda.client.jobhandling.result.ResultProcessorStrategy;
-import io.camunda.client.metrics.CamundaClientMetricsBridge;
-import io.camunda.client.metrics.MetricsRecorder;
-import java.time.Duration;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
