+++ clients/camunda-spring-boot-starter/src/test/java/io/camunda/client/jobhandling/parameter/DefaultParameterResolverStrategyTest.java
@@ -26,11 +26,9 @@ import io.camunda.client.api.response.ActivatedJob;
 import io.camunda.client.api.worker.JobClient;
 import io.camunda.client.bean.ParameterInfo;
 import io.camunda.client.impl.CamundaObjectMapper;
-import io.camunda.spring.client.annotation.Document;
-import io.camunda.spring.client.bean.ParameterInfo;
-import io.camunda.spring.client.jobhandling.DocumentContext;
-import io.camunda.spring.client.jobhandling.parameter.ParameterResolverStrategy.ParameterResolverStrategyContext;
 import io.camunda.client.jobhandling.DocumentContext;
+import io.camunda.client.jobhandling.parameter.ParameterResolverStrategy.ParameterResolverStrategyContext;
+import io.camunda.client.spring.bean.ParameterInfo;
 import io.camunda.zeebe.client.ZeebeClient;
 import java.util.List;
 import org.junit.jupiter.api.Test;
