+++ clients/camunda-spring-boot-starter/src/test/java/io/camunda/client/spring/config/AnnotationProcessorConfigurationTest.java
@@ -20,7 +20,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import io.camunda.client.CamundaClient;
 import io.camunda.client.bean.BeanInfo;
 import io.camunda.client.jobhandling.JobWorkerManager;
+import io.camunda.client.jobhandling.parameter.ParameterResolverStrategy;
+import io.camunda.client.jobhandling.result.ResultProcessorStrategy;
 import io.camunda.client.lifecycle.CamundaClientLifecycleAware;
+import io.camunda.client.metrics.MetricsRecorder;
 import io.camunda.client.spring.annotation.processor.AbstractCamundaAnnotationProcessor;
 import io.camunda.client.spring.configuration.AnnotationProcessorConfiguration;
 import io.camunda.client.spring.event.CamundaClientCreatedSpringEvent;
